fixed type in shapes

+++ src/shapes/Quadrilateral.java
@@ -1,10 +1,10 @@
 package shapes;
 
 abstract class Quadrilateral extends Shape implements Measurable{
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public Quadrilateral(int length, int width){
+    public Quadrilateral(double length, double width){
         this.length = length;
         this.width = width;
